# fix: implement MerchantMenu#getBukkitView to resolve trade crashes
## Description
PR addresses a `java.lang.NoSuchMethodError` occurring during villager trade interactions. The crash was caused by a missing implementation of `getBukkitView()` in `MerchantMenu`, which is required by newer mod loader environments.

By implementing a Mixin for `MerchantMenu`, we now properly initialize and return a `CraftMerchantView`, linking the merchant inventory to the Bukkit view and preventing the method-not-found exception.

## Changes
* **Mixin Implementation**: Added `getBukkitView()` to `MerchantMenu` via `ActualMerchantMenuMixin`.
* **View Synchronization**: Ensured `arclight$bukkitView` is lazily initialized and correctly references the bridge inventory.
* **Dependency Updates**: Updated project dependencies and imports to ensure compatibility with latest NMS and Bukkit mappings.

## Fixes
Fixes: https://github.com/IzzelAliz/Arclight/issues/2124

---
**Error Log Reference:**
`java.lang.NoSuchMethodError: 'org.bukkit.craftbukkit.v1_21_R1.inventory.view.CraftMerchantView net.minecraft.world.inventory.MerchantMenu.getBukkitView()'`
