+++ arclight-common/src/main/java/io/izzel/arclight/common/mixin/core/world/inventory/MerchantMenuMixin.java
@@ -1,86 +1,42 @@
 package io.izzel.arclight.common.mixin.core.world.inventory;
 
-import io.izzel.arclight.common.bridge.core.entity.EntityBridge;
-import io.izzel.arclight.common.bridge.core.world.IInventoryBridge;
-import net.minecraft.core.NonNullList;
-import net.minecraft.world.Container;
-import net.minecraft.world.entity.npc.AbstractVillager;
+import io.izzel.arclight.common.bridge.core.world.entity.player.PlayerBridge;
+import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.MerchantContainer;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.RecipeHolder;
+import net.minecraft.world.inventory.MerchantMenu;
 import net.minecraft.world.item.trading.Merchant;
-import org.bukkit.Location;
-import org.bukkit.craftbukkit.v.entity.CraftAbstractVillager;
-import org.bukkit.craftbukkit.v.entity.CraftHumanEntity;
-import org.bukkit.entity.HumanEntity;
-import org.bukkit.inventory.InventoryHolder;
+import org.bukkit.craftbukkit.v.inventory.CraftInventoryMerchant;
+import org.bukkit.craftbukkit.v.inventory.view.CraftMerchantView;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.ArrayList;
-import java.util.List;
+@Mixin(MerchantMenu.class)
+public abstract class MerchantMenuMixin extends AbstractContainerMenuMixin {
 
-@Mixin(MerchantContainer.class)
-public abstract class MerchantMenuMixin implements IInventoryBridge, Container {
+    @Shadow @Final private MerchantContainer tradeContainer;
 
-    // @formatter:off
-    @Shadow @Final private NonNullList<ItemStack> itemStacks;
-    @Shadow @Final private Merchant merchant;
-    // @formatter:on
+    private CraftMerchantView bukkitEntity;
+    private Inventory playerInventory;
+    private Merchant arclight$merchant;
 
-    private List<HumanEntity> transactions = new ArrayList<>();
-    private int maxStack = MAX_STACK;
-
-    @Override
-    public List<ItemStack> getContents() {
-        return this.itemStacks;
+    @Inject(method = "<init>(ILnet/minecraft/world/entity/player/Inventory;Lnet/minecraft/world/item/trading/Merchant;)V", at = @At("RETURN"))
+    private void arclight$init(int i, Inventory inventory, Merchant merchant, CallbackInfo ci) {
+        this.playerInventory = inventory;
+        this.arclight$merchant = merchant;
     }
 
     @Override
-    public void onOpen(CraftHumanEntity who) {
-        transactions.add(who);
-    }
+    public CraftMerchantView getBukkitView() {
+        if (bukkitEntity != null) {
+            return bukkitEntity;
+        }
 
-    @Override
-    public void onClose(CraftHumanEntity who) {
-        transactions.remove(who);
-        this.merchant.setTradingPlayer(null);
-    }
-
-    @Override
-    public List<HumanEntity> getViewers() {
-        return transactions;
-    }
-
-    @Override
-    public InventoryHolder getOwner() {
-        return this.merchant instanceof AbstractVillager ? ((CraftAbstractVillager) ((EntityBridge) this.merchant).bridge$getBukkitEntity()) : null;
-    }
-
-    @Override
-    public void setOwner(InventoryHolder owner) { }
-
-    @Override
-    public int getMaxStackSize() {
-        if (maxStack == 0) maxStack = MAX_STACK;
-        return this.maxStack;
-    }
-
-    @Override
-    public void setMaxStackSize(int size) {
-        this.maxStack = size;
-    }
-
-    @Override
-    public Location getLocation() {
-        return this.merchant instanceof AbstractVillager ? ((EntityBridge) this.merchant).bridge$getBukkitEntity().getLocation() : null;
-    }
-
-    @Override
-    public RecipeHolder<?> getCurrentRecipe() { return null; }
-
-    @Override
-    public void setCurrentRecipe(RecipeHolder<?> recipe) {
+        CraftInventoryMerchant inventory = new CraftInventoryMerchant(this.arclight$merchant, this.tradeContainer);
+        bukkitEntity = new CraftMerchantView(((PlayerBridge) this.playerInventory.player).bridge$getBukkitEntity(), inventory, (MerchantMenu) (Object) this, this.arclight$merchant);
+        return bukkitEntity;
     }
 }
